Light node and effect UI - Rename StringArray property type to TokenArray
And add an inspector for it so it can be used by CCLightNode too.

There is a parallel change in CCBReader but since this is just changing the
property name and not the type ID stored in the CCB files, this change can
be committed without that one and nothing will break.