Light node UI - Move light icons into their own layer
Add a CCNode to CocosScene and SceneGraph to act as the parent of all light icon sprites
(this is not the actual CCLightNode that is being edited, just a visual representation of it). This
prevents the light icons from scaling when the zoom is changed similar to the way joints are
drawn.