Light editing - Add a default light to the scene
If there are no lights in the scene already, use a default light that is centered in the stage.
This light is not saved with the scene, it's only a convenience for editing. If there are lights
in the scene but they're all disabled then the default light will not be used.