Light editing - Refactor lighting stuff out of AppDelegate
This addresses Nicky's feedback regarding the addition of even more code to AppDelegate which,
as he noted correctly, is already overly full of code from every other SB feature.  Instead of putting
lighting control code directly into AppDelegate, introduce the LightingHandler class to act as the
owner of lighting functionality. AppDelegate has a LightingHandler and calls to it when necessary
but otherwise it stays out of the way and lets LightingHandler do all the work.

In order to coordinate lighting visibility changes between LightingHandler and SequenceHandler,
this change introduces the LightVisibilityDelegate protocol and a light visibility notification. The
delegate protocol is for external actors like the SequenceHandler to tell the LightingHandler that
the visibility of a single light has changed. The notification goes in the other direction and is for the
LightingHandler to tell external actors that the visibility of all lights has changed.