This is a potential fix for issue #1242
The big slowdown we've been seeing recently seems to be due to image file preview
in menus (c4fee53f6e21d1e930177798fa01caf9778a8774) but I think the bigger problem
is that we appear to be completely rebuilding the inspector pane on every update of the
timeline position. Even if I comment out c4fee5 I still see a lot of heavyweight operations
happening on every timeline position change.

This is a proposed fix which does not rebuild the inspector with every update. Instead
it has the contained inspector values refresh themselves on every update. So far it
looks promising with big performance improvements. It looks like some inspector value
objects (InspectorColor3 for example) will need to have their refresh methods implemented
though since some controls are not refreshing in response to timeline changes.